+++ src/Classification/Classifier/Hosting/WhmcsClassifier.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Startwind\WebInsights\Classification\Classifier\Hosting;
+
+class WhmcsClassifier
+{
+    public const TAG = 'hosting:tools:whmcs';
+}